+++ templates/estimate_genome_size.sh
@@ -24,12 +24,20 @@ else
         ESTIMATED_GENOME_SIZE=`head -n1 ${OUTPUT}`
         if [ ${ESTIMATED_GENOME_SIZE} -gt "!{params.max_genome_size}" ]; then
             rm ${OUTPUT}
-            echo "!{sample} estimated genome size (${ESTIMATED_GENOME_SIZE} bp) exceeds the maximum allowed
-                  genome size (!{params.max_genome_size} bp). If this is unexpected, please investigate
-                  !{sample} to determine a cause (e.g. metagenomic, contaminants, etc...).
-                  Otherwise, adjust the --max_genome_size parameter to fit your need. Further
-                  analysis of !{sample} will be discontinued." | \
-            sed 's/^\s*//' > max-genome-size-depth-error.txt
+            echo "!{sample} estimated genome size (${ESTIMATED_GENOME_SIZE} bp) exceeds the maximum
+                  allowed genome size (!{params.max_genome_size} bp). If this is unexpected, please
+                  investigate !{sample} to determine a cause (e.g. metagenomic, contaminants, etc...).
+                  Otherwise, adjust the --max_genome_size parameter to fit your need. Further analysis
+                  of !{sample} will be discontinued." | \
+            sed 's/^\s*//' > !{sample}-genome-size-error.txt
+        elif [ ${ESTIMATED_GENOME_SIZE} -lt "!{params.min_genome_size}" ]; then
+            rm ${OUTPUT}
+            echo "!{sample} estimated genome size (${ESTIMATED_GENOME_SIZE} bp) is less than the minimum
+                  allowed genome size (!{params.min_genome_size} bp). If this is unexpected, please
+                  investigate !{sample} to determine a cause (e.g. metagenomic, contaminants, etc...).
+                  Otherwise, adjust the --min_genome_size parameter to fit your need. Further analysis
+                  of !{sample} will be discontinued." | \
+            sed 's/^\s*//' > !{sample}-genome-size-error.txt
         fi
     elif [ "!{params.genome_size}" == "min" ]; then
         # Use the minimum genome size based on completed RefSeq genomes